+++ R/CMap.R
@@ -100,8 +100,8 @@ prepareCMapZscores <- function(file, zscoresID=NULL) {
 #' Load matrix of CMap perturbation's differential expression z-scores
 #'
 #' @param data \code{perturbationChanges} object
-#' @param perturbationChanges Boolean: convert to \code{perturbationChanges}
-#'   object?
+#' @param inheritAttrs Boolean: convert to \code{perturbationChanges} object and
+#'   inherit attributes from \code{data}?
 #' @param verbose Boolean: print messages?
 #'
 #' @family functions related with the ranking of CMap perturbations
@@ -117,7 +117,7 @@ prepareCMapZscores <- function(file, zscoresID=NULL) {
 #'                                   "cmapGeneInfo.txt")
 #' zscores <- loadCMapZscores(perts[ , 1:10])
 #' }
-loadCMapZscores <- function(data, perturbationChanges=FALSE, verbose=TRUE) {
+loadCMapZscores <- function(data, inheritAttrs=FALSE, verbose=TRUE) {
     if (verbose) {
         msg <- paste("Loading CMap perturbation's differential expression",
                      "z-scores from %s...")
@@ -132,7 +132,7 @@ loadCMapZscores <- function(data, perturbationChanges=FALSE, verbose=TRUE) {
             zscores <- zscores[attr(data, "genes"), , drop=FALSE]
     }
 
-    if (perturbationChanges) {
+    if (inheritAttrs) {
         class(zscores) <- c("perturbationChanges", class(zscores))
         # Inherit input's attributes
         attrs <- attributes(data)
@@ -236,7 +236,7 @@ loadCMapGeneInfo <- function(file, nas) {
 loadCMapData <- function(file, type=c("metadata", "geneInfo", "zscores",
                                       "compoundInfo"),
                          zscoresID=NULL) {
-    if (is.null(file)) stop("File cannot be NULL, please provide a filename")
+    if (is.null(file)) stop("'file' cannot be NULL, please provide a filename")
 
     type <- match.arg(type)
     nas  <- c("", "NA", "na", "-666", "-666.0", "-666 -666",
@@ -359,7 +359,7 @@ filterCMapMetadata <- function(metadata, cellLine=NULL, timepoint=NULL,
 #' @param compoundInfo Data frame (CMap compound info) or character (respective
 #'   filepath to load data from file)
 #' @param loadZscores Boolean: load perturbation z-scores? Not recommended in
-#'   memory-constrained systems
+#'   systems with less than ~30GB available of RAM
 #'
 #' @importFrom R.utils gunzip
 #' @importFrom methods new
@@ -380,6 +380,7 @@ prepareCMapPerturbations <- function(metadata, zscores, geneInfo,
     if (is.character(geneInfo)) geneInfo <- loadCMapData(geneInfo, "geneInfo")
     if (is.character(zscores)) {
         zscores <- loadCMapData(zscores, "zscores", metadata$sig_id)
+        attr(zscores, "zscoresFilename") <- as.character(zscores)
     }
     if (is.character(compoundInfo)) {
         compoundInfo <- loadCMapData(compoundInfo, "compoundInfo")
@@ -396,8 +397,7 @@ prepareCMapPerturbations <- function(metadata, zscores, geneInfo,
     attr(zscores, "source") <- "CMap"
     attr(zscores, "type")   <- "perturbations"
 
-    if (loadZscores) zscores <- loadCMapZscores(zscores,
-                                                perturbationChanges=TRUE)
+    if (loadZscores) zscores <- loadCMapZscores(zscores, inheritAttrs=TRUE)
 
     # Display summary message of loaded perturbations
     filters <- attr(metadata, "filter")
@@ -534,16 +534,16 @@ calculateCellLineMean <- function(data, cellLine, metadata, rankPerCellLine) {
 #'
 #' # Rank similar CMap perturbations using only Spearman's correlation
 #' rankSimilarPerturbations(diffExprStat, perturbations, method="spearman")
-rankSimilarPerturbations <- function(diffExprGenes, perturbations,
+rankSimilarPerturbations <- function(input, perturbations,
                                      method=c("spearman", "pearson", "gsea"),
                                      geneSize=150, cellLineMean="auto",
                                      rankPerCellLine=FALSE) {
     metadata  <- attr(perturbations, "metadata")
     cellLines <- length(unique(metadata$cell_id))
     rankedPerts <- compareAgainstReference(
-        diffExprGenes, perturbations, method=method, geneSize=geneSize,
-        cellLines=cellLines, cellLineMean=cellLineMean,
-        rankByAscending=TRUE, rankPerCellLine=rankPerCellLine)
+        input, perturbations, method=method, geneSize=geneSize,
+        cellLines=cellLines, cellLineMean=cellLineMean, rankByAscending=TRUE,
+        rankPerCellLine=rankPerCellLine)
 
     # Relabel the "identifier" column name to be more descriptive
     pertType <- unique(metadata$pert_type)
@@ -569,17 +569,12 @@ rankSimilarPerturbations <- function(diffExprGenes, perturbations,
 #' Operations on a \code{perturbationChanges} object
 #'
 #' @param x \code{perturbationChanges} object
-#' @param ... Extra arguments passed to generic methods
+#' @param ... Extra arguments
 #' @param perturbation Character (perturbation identifier) or a
 #'   \code{similarPerturbations} table (from which the respective perturbation
 #'   identifiers are retrieved)
-#' @inheritParams prepareGSEApathways
-#' @param method Character: method to plot results (\code{spearman},
-#'   \code{pearson} or \code{gsea})
-#' @param genes Character: when plotting gene set enrichment analysis (GSEA),
-#'   plot top genes (\code{genes = "top"}), bottom genes
-#'   (\code{genes = "bottom"}) or both (\code{genes = "both"}); only used if
-#'   \code{method = "gsea"}
+#' @inheritParams compareAgainstReferencePerMethod
+#' @inheritParams plot.referenceComparison
 #'
 #' @importFrom methods is
 #' @importFrom stats setNames
@@ -612,31 +607,32 @@ rankSimilarPerturbations <- function(diffExprGenes, perturbations,
 #'
 #' # Currently unsupported!
 #' # plot(cmapPerturbationsCompounds, pert, diffExprStat, method="gsea")
-plot.perturbationChanges <- function(x, perturbation, diffExprGenes,
+plot.perturbationChanges <- function(x, perturbation, input,
                                      method=c("spearman", "pearson", "gsea"),
                                      geneSize=150,
                                      genes=c("both", "top", "bottom"), ...) {
     method <- match.arg(method)
 
-    if (is(perturbation, "similarPerturbations"))
+    if (is(perturbation, "similarPerturbations")) {
         perturbation <- perturbation[[1]]
+    }
 
     cellLinePerts <- colnames(x)[
         parseCMapID(colnames(x), cellLine=FALSE) %in% perturbation]
     isSummaryPert <- length(cellLinePerts) > 0
 
     if (length(perturbation) == 0) {
-        stop("One perturbation ID must be provided")
+        stop("a perturbation ID must be provided")
     } else if (length(perturbation) > 1) {
-        stop("Only one perturbation ID is currently supported")
+        stop("only one perturbation ID is currently supported")
     } else if (!perturbation %in% colnames(x) && !isSummaryPert) {
-        stop("Perturbation not found in the columns of the given dataset")
+        stop("perturbation not found in the columns of 'x'")
     }
 
     if (!isSummaryPert) cellLinePerts <- perturbation
     names(cellLinePerts) <- cellLinePerts
     if (is.character(x)) {
-        zscores <- loadCMapZscores(x[cellLinePerts])
+        zscores <- loadCMapZscores(x[cellLinePerts], verbose=FALSE)
     } else {
         zscores <- unclass(x)
     }
@@ -647,19 +643,14 @@ plot.perturbationChanges <- function(x, perturbation, diffExprGenes,
     }, zscores)
 
     if (method != "gsea") {
-        plotSingleCorr(data, perturbation, diffExprGenes, ...)
+        plotSingleCorr(data, perturbation, input)
     } else {
         if (length(data) > 1) {
-            stop("Plotting GSEA of multiple perturbations is currently ",
-                 "unsupported")
+            stop("plotting GSEA of multiple perturbations is not supported")
         }
         data <- unclass(data[[1]])
-        pathways <- prepareGSEApathways(diffExprGenes=diffExprGenes,
-                                        geneSize=geneSize)
-        genes <- match.arg(genes)
-        if (!any(genes %in% c("top", "both")))    pathways$top    <- NULL
-        if (!any(genes %in% c("bottom", "both"))) pathways$bottom <- NULL
-        plotGSEA(data, pathways$top, pathways$bottom, title=perturbation, ...)
+        geneset <- prepareGSEAgenesets(input, geneSize)
+        plotGSEA(data, geneset, genes, title=perturbation, ...)
     }
 }